refactor: allow for easy modification of calculations beyond addition
- add abstract class Operator
- make directories, and move them to look comfortable
