+++ README.md
@@ -22,12 +22,12 @@
 
 ## 리랙토링 기능 목록
 - 리팩토링은 객체지향의 관점에서 진행합니다. 리팩토링을 진행하다 기능 목록이 바뀔 수도 있습니다.
-- [ ] **분류**
+- [x] **분류**
   - [x] 문자열을 index를 기준으로 자르지 않고, 구분자를 기준으로 분류한다.
   - [x] 구분자를 객체화한다.(구분자와 관련된 요구사항이 변할 때, 유연하게 대처 가능)
 
-- [ ] **연산**
-  - [ ] 만약 덧셈이 아닌 다른 계산을 하게 된다면, 변경을 쉽게 할 수 있게 한다.
+- [x] **연산**
+  - [x] 만약 덧셈이 아닌 다른 계산을 하게 된다면, 변경을 쉽게 할 수 있게 한다.
 
 - [ ] **전역**
   - [ ] 구분자, 정규표현식을 지정할 수 있는 설정을 구현하고, 프로그램 시작 시에 설정될 수 있게 한다.
