+++ src/main/java/calculator/io/View.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.io;
 
 public class View {
