+++ src/main/java/calculator/operator/Operator.java
@@ -1,0 +1,5 @@
+package calculator.operator;
+
+public abstract class Operator {
+    abstract Double calculate();
+}
